[FIXED] use int for millisecond

+++ glp_clock.h
@@ -7,6 +7,7 @@ extern "C"
 #define game_loop_clock_h
 
 #include <stdbool.h>
+#include <stdint.h>
 
 struct glp_clock;
 
@@ -14,8 +15,10 @@ struct glp_clock* glp_clock_create();
 void glp_clock_release(struct glp_clock*);
 
 void glp_clock_start(struct glp_clock*);
-double glp_clock_get_time(struct glp_clock*);
-double glp_clock_get_during(struct glp_clock*, bool reset);
+
+// return in millisecond
+uint32_t glp_clock_get_time(struct glp_clock*);
+uint32_t glp_clock_get_during(struct glp_clock*, bool reset);
 
 #endif // game_loop_clock_h
